pyproject.toml
[project]
name = "agno-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno==1.8.0",
    "fastapi>=0.117.1",
    "google-genai>=1.38.0",
    "groq>=0.31.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "tavily-python>=0.7.12",
    "uvicorn>=0.36.0",
]
